+++ database/queries/init/USER.sql
@@ -2,7 +2,7 @@ CREATE TABLE gro."ETH_BLOCKS" (
     block_number INTEGER NOT NULL,
     block_timestamp INTEGER NULL,
     block_date TIMESTAMP (6) NULL,
-    network_id SMALLINT NULL,
+    network_id INTEGER NULL,
     creation_date TIMESTAMP (6) NULL,
     CONSTRAINT "ETH_BLOCKS_pkey" PRIMARY KEY (block_number) 
         NOT DEFERRABLE INITIALLY IMMEDIATE
@@ -12,7 +12,7 @@ ALTER TABLE gro."ETH_BLOCKS" OWNER to postgres;
 
 CREATE TABLE gro."SYS_USER_LOADS" (
     table_name CHARACTER VARYING (50) NOT NULL,
-    network_id SMALLINT NULL,
+    network_id INTEGER NULL,
     target_date TIMESTAMP (6) NULL,
     records_loaded INTEGER NULL,
     creation_date TIMESTAMP (6) NULL
@@ -22,7 +22,7 @@ ALTER TABLE gro."SYS_USER_LOADS" OWNER to postgres;
 
 CREATE TABLE gro."TOKEN_PRICE" (
    price_date       TIMESTAMP (6) NOT NULL,
-   network_id       SMALLINT NULL,
+   network_id       INTEGER NULL,
    gvt_value        NUMERIC (20, 8) NULL,
    pwrd_value       NUMERIC (20, 8) NULL,
    gro_value        NUMERIC (20, 8) NULL,
@@ -40,7 +40,7 @@ CREATE TABLE gro."USER_STD_FACT_APPROVALS" (
     block_number INTEGER NOT NULL,
     block_timestamp INTEGER NULL,
     approval_date TIMESTAMP (6) NULL,
-    network_id SMALLINT NULL,
+    network_id INTEGER NULL,
     stablecoin_id SMALLINT NULL,
     tx_hash CHARACTER VARYING (66) NOT NULL,
     sender_address CHARACTER VARYING (42) NOT NULL,
@@ -57,7 +57,7 @@ ALTER TABLE gro."USER_STD_FACT_APPROVALS" OWNER to postgres;
 
 CREATE TABLE gro."USER_STD_TMP_APPROVALS" (
     block_number INTEGER NOT NULL,
-    network_id SMALLINT NULL,
+    network_id INTEGER NULL,
     stablecoin_id SMALLINT NULL,
     tx_hash CHARACTER VARYING (66) NOT NULL,
     sender_address CHARACTER VARYING (42) NOT NULL,
@@ -77,7 +77,7 @@ CREATE TABLE gro."USER_STD_FACT_TRANSFERS" (
     block_timestamp INTEGER NULL,
     transfer_date TIMESTAMP (6) NULL,
     tx_hash CHARACTER VARYING (66) NOT NULL,
-    network_id SMALLINT NULL,
+    network_id INTEGER NULL,
     transfer_type CHARACTER VARYING (20) NOT NULL,
     user_address CHARACTER VARYING (42) NOT NULL,
     referral_address CHARACTER VARYING (42) NULL,
@@ -108,7 +108,7 @@ ALTER TABLE gro."USER_STD_FACT_TRANSFERS" OWNER to postgres;
 CREATE TABLE gro."USER_STD_TMP_DEPOSITS" (
     block_number INTEGER NOT NULL,
     tx_hash CHARACTER VARYING (66) NOT NULL,
-    network_id SMALLINT NULL,
+    network_id INTEGER NULL,
     transfer_type CHARACTER VARYING (20) NULL,
     user_address CHARACTER VARYING (42) NOT NULL,
     referral_address CHARACTER VARYING (42) NULL,
@@ -130,7 +130,7 @@ ALTER TABLE gro."USER_STD_TMP_DEPOSITS" OWNER to postgres;
 CREATE TABLE gro."USER_STD_TMP_WITHDRAWALS" (
     block_number INTEGER NOT NULL,
     tx_hash CHARACTER VARYING (66) NOT NULL,
-    network_id SMALLINT NULL,
+    network_id INTEGER NULL,
     transfer_type CHARACTER VARYING (20) NULL,
     user_address CHARACTER VARYING (42) NOT NULL,
     referral_address CHARACTER VARYING (42) NULL,
@@ -154,7 +154,7 @@ ALTER TABLE gro."USER_STD_TMP_WITHDRAWALS" OWNER to postgres;
 
 -- CREATE TABLE gro."USER_STD_FACT_BALANCES_OLD" (
 --     balance_date TIMESTAMP (6) NOT NULL,
---     network_id SMALLINT NULL,
+--     network_id INTEGER NULL,
 --     user_address CHARACTER VARYING (42) NOT NULL,
 --     usd_value NUMERIC (20, 8) NULL,
 --     pwrd_value NUMERIC (20, 8) NULL,
@@ -167,7 +167,7 @@ ALTER TABLE gro."USER_STD_TMP_WITHDRAWALS" OWNER to postgres;
 
 CREATE TABLE gro."USER_STD_FACT_BALANCES" (
     balance_date     TIMESTAMP (6) NOT NULL,
-    network_id       SMALLINT NULL,
+    network_id       INTEGER NULL,
     user_address     CHARACTER VARYING (42) NOT NULL,
     gvt_unstaked_amount NUMERIC (20, 8) NULL,
     pwrd_unstaked_amount NUMERIC (20, 8) NULL,
@@ -203,7 +203,7 @@ ALTER TABLE gro."USER_STD_FACT_BALANCES" OWNER to postgres;
 -- Same as USER_STD_FACT_BALANCES: intended to do on-demand extractions at a specific date & time
 CREATE TABLE gro."USER_STD_FACT_BALANCES_SNAPSHOT" (
     balance_date TIMESTAMP (6) NOT NULL,
-    network_id SMALLINT NULL,
+    network_id INTEGER NULL,
     user_address CHARACTER VARYING (42) NOT NULL,
     gvt_unstaked_amount NUMERIC (20, 8) NULL,
     pwrd_unstaked_amount NUMERIC (20, 8) NULL,
@@ -238,7 +238,7 @@ ALTER TABLE gro."USER_STD_FACT_BALANCES_SNAPSHOT" OWNER to postgres;
 
 -- CREATE TABLE gro."USER_STD_FACT_NET_RETURNS_OLD" (
 --    balance_date        TIMESTAMP (6) NOT NULL,
---    network_id          SMALLINT NULL,
+--    network_id          INTEGER NULL,
 --    user_address        CHARACTER VARYING (42) NOT NULL,
 --    total_value         NUMERIC (20, 8) NULL,
 --    pwrd_value          NUMERIC (20, 8) NULL,
@@ -253,7 +253,7 @@ ALTER TABLE gro."USER_STD_FACT_BALANCES_SNAPSHOT" OWNER to postgres;
 
 CREATE TABLE gro."USER_STD_FACT_NET_RETURNS" (
     balance_date TIMESTAMP (6) NOT NULL,
-    network_id SMALLINT NULL,
+    network_id INTEGER NULL,
     user_address CHARACTER VARYING (42) NOT NULL,
     total_unstaked_value NUMERIC (20, 8) NULL,
     pwrd_unstaked_value NUMERIC (20, 8) NULL,
